--- cmd/httpx/banner.go
@@ -1,24 +1,0 @@
-package main
-
-import "github.com/projectdiscovery/gologger"
-
-const banner = `
-    __    __  __       _  __
-   / /_  / /_/ /_____ | |/ /
-  / __ \/ __/ __/ __ \|   / 
- / / / / /_/ /_/ /_/ /   |  
-/_/ /_/\__/\__/ .___/_/|_|  
-             /_/            
-`
-
-// Version is the current version of httpx
-const Version = `1.0.2`
-
-// showBanner is used to show the banner to the user
-func showBanner() {
-	gologger.Printf("%s\n", banner)
-	gologger.Printf("\t\tprojectdiscovery.io\n\n")
-
-	gologger.Labelf("Use with caution. You are responsible for your actions\n")
-	gologger.Labelf("Developers assume no liability and are not responsible for any misuse or damage.\n")
-}
